purely-fresh-backend/requirements.txt
bcrypt==4.1.2
email-validator==2.1.0.post1
fastapi==0.105.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
jose==1.0.0
openai==0.27.8
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose==3.3.0
python-jwt==4.1.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
uvicorn==0.24.0
python-dotenv  
